==> notebooks/01_data_cleaning.ipynb <==
# Load Data

import pandas as pd

%cd ..
from src.data.clean_data import drop_columns, drop_all_null_rows, drop_null_rows, filter_dataframe
%cd -

annual_collision_report_raw = pd.read_csv("../data/raw/Annual_Collision_Report.csv")
average_annual_weekday_traffic_volume_raw = pd.read_csv("../data/raw/Average_Annual_Weekday_Traffic_Volumes.csv")
weather_data_hourly = pd.read_csv("../data/raw/Weather_Data_Hourly.csv", low_memory=False)

# Processing

to_drop_collision = [ "Population", "Private Passenger Vehicles", "Private Motorcycles", "Collisions per 1,000 Population", "Property Damage Only (PDO) Collisions", "Injuries+Fatalities per 1,000 Population", "Injuries per 1,000 Population", "Fatalities per 1,000 Population", "FATALITY: BICYCLIST", "FATALITY: MOTORCYCLIST", "FATALITY: PEDESTRIAN", "FATALITY: VEHICLE DRIVER", "FATALITY: VEHICLE PASSENGER", "SERIOUS INJURIES: BICYCLIST", "MINOR INJURIES: BICYCLIST", "SERIOUS INJURY: MOTORCYCLIST", "MINOR INJURIES: MOTORCYCLIST", "SERIOUS INJURIES: PEDESTRIAN", "MINOR INJURIES: PEDESTRIAN", "Total Serious Injuries", "Total Minor Injuries", "Total Serious + Minor Injuries", "Total Fatalities + Injuries", "FATALITIES: INTERSECTION PERCENT", "FATALITIES: MIDBLOCK PERCENT", "INJURIES: INTERSECTION PERCENT", "INJURIES: MIDBLOCK PERCENT" ]
to_drop_volume = traffic_volume_drop = [ "Location", "Geometry Point", "Latitude", "Longitude"]
to_drop_weather = [ "Row ID", "Station Province", "Location", "Station Elevation (m)", "Station Climate Identifier", "Station WMO Identifier", "Station TC Identifier", "Station Note", "Dewpoint Temperature (degrees C)", "Relative Humidity", "Wind Direction (10s degrees)", "Air Pressure (kPa)", "Humidex", "Wind Chill (degrees C)", "Station Latitude", "Station Longitude"]


collision_new = drop_columns(annual_collision_report_raw, to_drop_collision)
traffic_new = drop_columns(average_annual_weekday_traffic_volume_raw, to_drop_volume)
weather_new = drop_columns(weather_data_hourly, to_drop_weather)

print("Traffic Volume Dataset NULL values: \n" + str(traffic_new.isnull().sum()))
traffic_new[traffic_new.isnull()]

traffic_new = drop_all_null_rows(traffic_new)
weather_new = drop_all_null_rows(weather_new)

traffic_new = drop_null_rows(traffic_new, ['Average Daily Volume'])
weather_new = drop_null_rows(weather_new, ["Temperature (degrees C)", "Wind Speed (km/h)", "Visibility (km)"])

edmonton_weather_data = filter_dataframe(weather_new, "Station Name", "edmonton")
edmonton_weather_data

# Droping Meta Columns

edmonton_weather_data = drop_columns(edmonton_weather_data, ["Station Name"])

# Save To CSV:

collision_new.to_csv("../data/processed/traffic_collisions.csv")
traffic_new.to_csv("../data/processed/traffic_volume.csv")
edmonton_weather_data.to_csv("../data/processed/edmonton_weather.csv")
