==> notebooks/02_feature_engineering.ipynb <==
import pandas as pd

# Loading Processed Data

weather = pd.read_csv("../data/processed/edmonton_weather.csv")
collisions = pd.read_csv("../data/processed/traffic_collisions.csv")
volume = pd.read_csv("../data/processed/traffic_volume.csv")

weather

collisions.drop(["Unnamed: 0"], axis=1, inplace=True)
collisions

volume

# Weather Dataset

#Removing Column
weather.drop(["Unnamed: 0"], axis=1, inplace=True)

# Date and Time columns:
weather["Date and Time"] = pd.to_datetime(weather["Date and Time"])
weather

#Unique Years
years = sorted(list(weather["Date and Time"].dt.year.unique()))[10:24] #Matching with Dataset
years

#Getting Average Values for each year

dataframes = []
temp_average = []
wind_speed_average = []
visibility_average = []

for year in years:
    df_year = weather[weather["Date and Time"].dt.year == year]
    dataframes.append(df_year)

for df in dataframes:
    temp_avg = df["Temperature (degrees C)"].mean()
    wind_speed_avg = df["Wind Speed (km/h)"].mean()
    visibility_avg = df["Visibility (km)"].mean()
    
    temp_average.append(temp_avg)
    wind_speed_average.append(wind_speed_avg)
    visibility_average.append(visibility_avg)

#Making new Data Frame

yearly_average_weather = pd.DataFrame({
    "Year": years,
    "Average temperature (degrees C)": temp_average,
    "Average wind Speed (km/h)": wind_speed_average,
    "Average visibility (km)": visibility_average
})

yearly_average_weather

# Volume Dataset

#Removing Columns
volume.drop(["Unnamed: 0"], axis=1, inplace=True)
volume

dataframes = []
yearly_average__avg_daily_volume = []

for year in years:
    df_year = volume[volume["Year"] == year]
    dataframes.append(df_year)

#No Data from Year = 2010 and 2023 (i = 0 and 13)

for df in dataframes:
    avg_volume = df["Average Daily Volume"].mean()
    yearly_average__avg_daily_volume.append(avg_volume)

yearly_average_daily_volumes = pd.DataFrame({
    "Year": years,
    "Yearly Average of Daily Volume": yearly_average__avg_daily_volume
})

yearly_average_daily_volumes

# Merge:

merged = yearly_average_weather.merge(yearly_average_daily_volumes, on="Year")
merged = merged.merge(collisions, on="Year")
merged

# Injury Share Feature

merged["RATIO FATALITIES + INJURIES"] = merged["TOTAL FATAL & INJURY COLLISIONS"] / merged["Total Collisions"]
merged

# Save to CSV

merged.to_csv("../data/processed/merged_features.csv")
